+++ src/layer/cesium/vectorCesiumImpl.ts
@@ -1,12 +1,5 @@
-import { unByKey } from 'ol/Observable.js';
-import {
-  PrimitiveCollection,
-  type Scene,
-  type SplitDirection,
-} from '@vcmap-cesium/engine';
+import { PrimitiveCollection, type SplitDirection } from '@vcmap-cesium/engine';
 import type VectorSource from 'ol/source/Vector.js';
-import type { EventsKey } from 'ol/events.js';
-import type { Feature } from 'ol/index.js';
 import { vcsLayerName } from '../layerSymbols.js';
 import LayerImplementation from '../layerImplementation.js';
 import { synchronizeFeatureVisibilityWithSource } from '../vectorHelpers.js';
@@ -18,6 +11,10 @@ import type StyleItem from '../../style/styleItem.js';
 import type FeatureVisibility from '../featureVisibility.js';
 import type GlobalHider from '../globalHider.js';
 import VectorContext from './vectorContext.js';
+import {
+  createSourceVectorContextSync,
+  SourceVectorContextSync,
+} from './sourceVectorContextSync.js';
 
 /**
  * represents a specific vector layer for cesium.
@@ -42,21 +39,11 @@ class VectorCesiumImpl
 
   private _featureVisibilityListeners: (() => void)[] = [];
 
-  // eslint-disable-next-line class-methods-use-this
-  private _removeVectorPropertiesChangeHandler = (): void => {};
+  private _rootCollection: PrimitiveCollection;
 
-  protected _rootCollection: PrimitiveCollection;
+  private _context: VectorContext | null = null;
 
-  private _olListeners: (EventsKey | EventsKey[])[] = [];
-
-  /**
-   * A set of ol.Features to add once the map is back to cesium
-   */
-  private _featureToAdd: Set<Feature> = new Set();
-
-  protected _context: VectorContext | null = null;
-
-  private _scene: Scene | undefined = undefined;
+  private _sourceVectorContextSync: SourceVectorContextSync | undefined;
 
   globalHider: GlobalHider | undefined;
 
@@ -74,52 +61,21 @@ class VectorCesiumImpl
     this.globalHider = options.globalHider;
   }
 
-  private _addListeners(): void {
-    this._olListeners.push(
-      this.source.on('addfeature', (event) => {
-        this._addFeature(event.feature as Feature).catch(() => {
-          this.getLogger().error('failed to convert feature');
-        });
-      }),
-    );
-
-    this._olListeners.push(
-      this.source.on('removefeature', (event) => {
-        this._removeFeature(event.feature as Feature);
-      }),
-    );
-
-    this._olListeners.push(
-      this.source.on('changefeature', (event) => {
-        this._featureChanged(event.feature as Feature).catch((_e) => {
-          this.getLogger().error('failed to convert feature');
-        });
-      }),
-    );
-
-    this._removeVectorPropertiesChangeHandler =
-      this.vectorProperties.propertyChanged.addEventListener(
-        this.refresh.bind(this),
-      );
-  }
-
-  protected _setupContext(cesiumMap: CesiumMap): Promise<void> {
-    const rootCollection = this._rootCollection;
-    this._context = new VectorContext(
-      cesiumMap,
-      rootCollection,
-      this.splitDirection,
-    );
-    cesiumMap.addPrimitiveCollection(rootCollection);
-    return Promise.resolve();
-  }
-
   async initialize(): Promise<void> {
     if (!this.initialized) {
-      this._scene = this.map.getScene();
-      this._addListeners();
-      this._addFeatures(this.source.getFeatures());
-      await this._setupContext(this.map);
+      this._context = new VectorContext(
+        this.map,
+        this._rootCollection,
+        this.splitDirection,
+      );
+      this.map.addPrimitiveCollection(this._rootCollection);
+      this._sourceVectorContextSync = createSourceVectorContextSync(
+        this.source,
+        this._context,
+        this.map.getScene()!,
+        this.style.style,
+        this.vectorProperties,
+      );
     }
     await super.initialize();
     if (this.splitDirection) {
@@ -127,62 +83,18 @@ class VectorCesiumImpl
     }
   }
 
-  private _addFeatures(features: Feature[]): void {
-    // TODO we should make this non-blocking to better handle larger data sets check in RIWA Impl
-    features.forEach((f) => {
-      this._addFeature(f).catch((err) => {
-        this.getLogger().error('failed to convert feature', f, err);
-      });
-    });
-  }
-
-  /**
-   * converts a feature and adds the associated primitives to the collection of primitives
-   */
-  private async _addFeature(feature: Feature): Promise<void> {
-    if (this.active) {
-      // XXX cluster check here? or on init?
-      await this._context!.addFeature(
-        feature,
-        this.style.style,
-        this.vectorProperties,
-        this._scene as Scene,
-      );
-    } else {
-      this._featureToAdd.add(feature);
-    }
-  }
-
   /**
    * Forces a complete re-render of all features.
    */
   refresh(): void {
-    this._context?.clear();
-    this._addFeatures(this.source.getFeatures());
-  }
-
-  /**
-   * removes the primitive of the specified feature
-   */
-  private _removeFeature(feature: Feature): void {
-    this._context?.removeFeature(feature);
-    this._featureToAdd.delete(feature);
-  }
-
-  /**
-   * called when a features property have changed
-   */
-  private async _featureChanged(feature: Feature): Promise<void> {
-    this._featureToAdd.delete(feature);
-    await this._addFeature(feature);
+    this._sourceVectorContextSync?.refresh();
   }
 
   async activate(): Promise<void> {
     if (!this.active) {
       await super.activate();
       if (this.active) {
-        this._addFeatures([...this._featureToAdd]);
-        this._featureToAdd.clear();
+        this._sourceVectorContextSync?.activate();
         this._rootCollection.show = true;
         if (this._featureVisibilityListeners.length === 0) {
           this._featureVisibilityListeners =
@@ -198,6 +110,7 @@ class VectorCesiumImpl
 
   deactivate(): void {
     super.deactivate();
+    this._sourceVectorContextSync?.deactivate();
     this._rootCollection.show = false;
     this._featureVisibilityListeners.forEach((cb) => {
       cb();
@@ -208,11 +121,10 @@ class VectorCesiumImpl
   updateStyle(style: StyleItem, silent?: boolean): void {
     this.style = style;
     if (this.initialized && !silent) {
-      const features = this.source.getFeatures().filter((f) => !f.getStyle());
-      features.forEach((f) => {
-        // eslint-disable-next-line no-void
-        void this._featureChanged(f);
-      });
+      this.source
+        .getFeatures()
+        .filter((f) => !f.getStyle())
+        .forEach((f) => f.changed());
     }
   }
 
@@ -223,43 +135,17 @@ class VectorCesiumImpl
     }
   }
 
-  protected _destroyCollection(): void {
-    this.map.removePrimitiveCollection(this._rootCollection);
-  }
-
   destroy(): void {
     if (this.initialized) {
+      this._sourceVectorContextSync?.destroy();
       this._context?.destroy();
-      this._destroyCollection();
+      this.map.removePrimitiveCollection(this._rootCollection);
     }
     this._context = null;
-    this._scene = undefined;
-    this._removeVectorPropertiesChangeHandler();
-    this._olListeners.forEach((listener) => {
-      unByKey(listener);
-    });
-    this._olListeners = [];
-    this._featureToAdd.clear();
     this._featureVisibilityListeners.forEach((cb) => {
       cb();
     });
     this._featureVisibilityListeners = [];
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    this.source = null;
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    this.vectorProperties = null;
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    this.featureVisibility = null;
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    this.style = null;
-    this.globalHider = undefined;
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
-    this._rootCollection = null;
     super.destroy();
   }
 }